fix: use /admin/returns/:id/receive for swap returns
