+++ packages/medusa-js/src/resources/admin/orders.ts
@@ -1,22 +1,21 @@
 import {
-  AdminPostOrdersOrderReq,
-  AdminOrdersRes,
   AdminGetOrdersParams,
   AdminOrdersListRes,
-  AdminPostOrdersReq,
-  AdminPostOrdersOrderRefundsReq,
+  AdminOrdersRes,
+  AdminPostOrdersOrderClaimsClaimFulfillmentsReq,
+  AdminPostOrdersOrderClaimsClaimReq,
+  AdminPostOrdersOrderClaimsClaimShipmentsReq,
+  AdminPostOrdersOrderClaimsReq,
   AdminPostOrdersOrderFulfillmentsReq,
-  AdminPostOrdersOrderShipmentReq,
+  AdminPostOrdersOrderRefundsReq,
+  AdminPostOrdersOrderReq,
   AdminPostOrdersOrderReturnsReq,
+  AdminPostOrdersOrderShipmentReq,
   AdminPostOrdersOrderShippingMethodsReq,
   AdminPostOrdersOrderSwapsReq,
-  AdminPostOrdersOrderSwapsSwapReceiveReq,
   AdminPostOrdersOrderSwapsSwapFulfillmentsReq,
   AdminPostOrdersOrderSwapsSwapShipmentsReq,
-  AdminPostOrdersOrderClaimsReq,
-  AdminPostOrdersOrderClaimsClaimFulfillmentsReq,
-  AdminPostOrdersOrderClaimsClaimShipmentsReq,
-  AdminPostOrdersOrderClaimsClaimReq,
+  AdminPostOrdersReq,
 } from "@medusajs/medusa"
 import qs from "qs"
 import { ResponsePromise } from "../../typings"
@@ -149,15 +148,6 @@ class AdminOrdersResource extends BaseResource {
   cancelSwap(id: string, swapId: string): ResponsePromise<AdminOrdersRes> {
     const path = `/admin/orders/${id}/swaps/${swapId}/cancel`
     return this.client.request("POST", path)
-  }
-
-  receiveSwap(
-    id: string,
-    swapId: string,
-    payload: AdminPostOrdersOrderSwapsSwapReceiveReq
-  ): ResponsePromise<AdminOrdersRes> {
-    const path = `/admin/orders/${id}/swaps/${swapId}/receive`
-    return this.client.request("POST", path, payload)
   }
 
   fulfillSwap(
