+++ packages/medusa-js/src/resources/admin/returns.ts
@@ -1,5 +1,6 @@
 import {
   AdminGetReturnsParams,
+  AdminPostReturnsReturnReceiveReq,
   AdminReturnsCancelRes,
   AdminReturnsListRes,
   AdminReturnsRes,
@@ -22,11 +23,15 @@ class AdminReturnsResource extends BaseResource {
   /**
    * @description receive a return
    * @param id id of the return to receive.
+   * @param payload items to receive and an optional refund amount
    * @returns the return
    */
-  receive(id: string): ResponsePromise<AdminReturnsRes> {
+  receive(
+    id: string,
+    payload: AdminPostReturnsReturnReceiveReq
+  ): ResponsePromise<AdminReturnsRes> {
     const path = `/admin/returns/${id}/receive`
-    return this.client.request("POST", path)
+    return this.client.request("POST", path, payload)
   }
 
   /**
